Update my local forked branch

+++ cppu/source/uno/destr.hxx
@@ -33,7 +33,6 @@ namespace cppu
 //#### destruction #################################################################################
 
 
-
 void destructStruct(
     void * pValue,
     typelib_CompoundTypeDescription * pTypeDescr,
@@ -133,7 +132,7 @@ inline void _destructAny(
         break;
     }
 #if OSL_DEBUG_LEVEL > 0
-    pAny->pData = reinterpret_cast<void *>(0xdeadbeef);
+    pAny->pData = reinterpret_cast<void *>((uintptr_t)0xdeadbeef);
 #endif
 
     ::typelib_typedescriptionreference_release( pType );
